Moved DAO intialization from DAOPMLogger to PMDAO
